Export the data required to implement interactivity for charts.
We reach into the chart server while it's rendering, at the points where
it has access to various data that we need (x/y coordinates etc), and
store the information as data attributes on the nearest available SVG
nodes.

There is a pile of code in the figure.nz codebase which reads this
information out and builds the interactivity with it.